requirements.txt
langchain_core
langchain_openai
streamlit
langchain_community
python-dotenv
langchain
langchain-huggingface
langchain-groq
langchain-google-genai
langserve
fastapi
uvicorn
sse_starlette